==> 5_Total_modelling_LazyPredict_2.ipynb <==
!pip install git+https://github.com/shankarpandala/lazypredict.git

!pip install lazypredict

from lazypredict.Supervised import LazyClassifier
from lazypredict.Supervised import LazyRegressor
from sklearn.model_selection import train_test_split
import os
import pandas as pd
import numpy as np

from google.colab import files
uploaded = files.upload()
import io
df = pd.read_csv(io.BytesIO(uploaded['churn-data_10_DFS_pearsons-selected.csv']))

data_1 = df.drop(df.columns[0], axis=1)
data_1.head(3)

data_1.shape

# random sampling (20%)
data = data_1.sample(frac =.2)
data.shape

X = data.iloc[:,1:620].values
y = data.loc[:,'churn'].values
print(X.shape, y.shape)

X_train, X_test, y_train, y_test = train_test_split(X, y,
                                                    test_size=.2,random_state =123)

print(X_train.shape, y_train.shape,
      X_test.shape, y_test.shape)

clf = LazyClassifier(verbose=0,ignore_warnings=True, custom_metric=None)
models,predictions = clf.fit(X_train, X_test, y_train, y_test)

print(models)

reg = LazyRegressor(verbose=0, ignore_warnings=False, custom_metric=None)
models, predictions = reg.fit(X_train, X_test, y_train, y_test)

print(models)
