==> 6_Support_vector_classifier_2.ipynb <==
import numpy as np
import pandas as pd
import matplotlib.pyplot as plt
import seaborn as sns
%matplotlib inline
from sklearn.model_selection import train_test_split

from google.colab import files
uploaded = files.upload()
import io
df = pd.read_csv(io.BytesIO(uploaded['churn-data_5_encoded.csv']))

data = df.drop(df.columns[0], axis=1)
data = data.dropna()
data.isnull().sum()

X = data.iloc[:,:36].values
Y = data.loc[:,'churn'].values
print(X.shape, Y.shape)

train_x, test_x, train_y, test_y = train_test_split(
    X, Y, test_size=0.2, random_state=65
)
print(train_x.shape, train_y.shape,
      test_x.shape, test_y.shape)

# model training
from sklearn.svm import SVC

SVC_model= SVC()
SVC_model.fit(train_x, train_y)

y_pred = SVC_model.predict(test_x)

# model testing
from sklearn.metrics import classification_report, confusion_matrix

confusion_matrix(test_y, y_pred)

print(classification_report(test_y, y_pred))

# tuning

from sklearn.model_selection import GridSearchCV

svm = SVC()
grid_parameters = {'C':[0.01,0.1,1, 10, 100, 1000],'gamma':[1, 0.1, 0.01, 0.001, 0.0001],
                   'kernel':['rbf', 'poly', 'sigmoid']}
grid_search = GridSearchCV(svm,grid_parameters, cv=5, verbose=2)

#grid_parameters = {'C':[0.1,1, 10],'gamma':[1, 0.1, 0.01],
#                   'kernel':['rbf']}                
#grid_search = GridSearchCV(svm,grid_parameters, cv=5 , verbose=2)

grid_search.fit(train_x, train_y)

grid_search.best_estimator_

grid_search.best_params_

y_pred_grid= grid_search.predict(test_x)

confusion_matrix(test_y, y_pred_grid)

print(classification_report(test_y, y_pred_grid))
